(Change Version)
Added fall through /etc/lsb-release test to update distro data tests.

This adds one more failsafe test to future proof the stuff against
id failures. 

I tried to also use the standard lsb_release function but for some reason
when it's run inside inxi and sent to awk it gets incredibly slowed down.
This is odd, since that command usually runs quite fast in shell.

But it's too slow as is for now to use, it makes everything hang.

Maybe later I'll add an option to trigger it or something.