(Change Version) This is another milestone, hit much sooner than we thought. We have now restored the broken konversation native script support.
This will be of special interest to distros who want to include inxi as native, or to packagers.

Basically, all that is needed is to set this symbolic link:

ln -s <inxi location, ie: /usr/bin/inxi, or /usr/local/bin/inxi> /var/share/apps/konversation/scripts/inxi

in the package post install script.

This will automatically enable native, ie: /inxi support in konversation.

Fixed a few other small bugs, and cleaned up code and comments as well.

We were almost going to call this release 1.0.0, but we'll leave that until it's absolutely certain that everything is 
debugged and working.