version upgrade got rid of that horrible banlist array/function thing, now it's hard coded.
On a dual core system, this saves literally 0.1 seconds on execution, about 10-20% of execution time.

this also fixes a bug that has always existed, which is if you use spaces to replace instead of ''.

Another advantage of this new method it fixes the fact that each and every string character would be considered a match, null replaced with null.

This was due to the fact that each banlist string started with |, ie, it contained a null character in the search pattern... sigh.

This is why we don't use stupid bash tricks, ok people? They  hide bugs, and make the code write only.