totally synced separators in output, now if in console, uses : universally, or if color scheme is 0.
This is set dynamically as well. For irc it uses no : except for line starters and item starters, like:
ID: or IF: or 1: and so on. Also made cpu per core clock speed the same, using now 1: instead of (1)

that saves one character too per core, but makes it all the same, consistent.

This also now becomes a user adjustable parameter, via config files, can set irc/console separator character(s)

Ok, bored, now on with the day.