added feature where a different ip page can be added by whoever.
That would need to be set in the inxi override config file. The page must show only
one line of output, and that last item on that line must be the ip address.

Ok, that's it, not a big deal, but some people may prefer a non smxi ip address page
for whatever reason, so have at it. Or not.