small change, with -xx, now inxi shows one driver per line, but shows the total disk space on a separate line
that makes it a bit more readable. New version number, new tarball.

I debated making it one per line default as soon as -x is triggered but will leave it this way for now, with -xx
you get the easier to read single line output in all cases, so that's probably good enough.