New version, new tarball. Improved B_ALLOW_UPDATES handling, now if set to false, turns off all -h and -H menu options for updating. Also triggers an error message if you use -U or -! <10-16/http://>.
Distro maintainers, take note, if you used the B_ALLOW_UPDATES flag, you no longer
need to change the code anywhere, the error messages and blocking the -h output for
update features is automatic as soon as the flag is set to 'false'.

I needed to change the -! handling because -! is now also being used for extra features 
like -! 31 and -! 32 and probably more stuff in the future, plus the -! 30 used by 
things like the inxi gui tool being worked on by trash80.

Also included in this version are more bsd changes, including initial function for pciconf
data parsing, this will be used for -A, -G, and -N options for card data.

Further bsd improvements are better error/no data available messages for -D and -o.