New version, new tarball. Fixed a bug in Xorg where it shows drivers as unloaded when they are actually loaded. Since we can't fix xorg, inxi will try to work around this bug by validating one step further in the Xorg.0.log data, to confirm that drivers noted as loaded/unloaded/failed are actually running the display(s) of the system.
There is a possible case of error that might happen due to this change in the case of a system with
a complex xorg that uses two drivers/modules to run two different displays, ie, nvidia on one, and amd
on the other, for example, or intel/nvidia, etc. However, if that bug appears, we'll get that data set
of debugging output and fix it at that point. 

This fix repairs an existing xorg bug that is unlikely to get fixed any time soon (the call to load the
detected drivers, eg, vesa, intel, is repeated, causing a failure of driver already loaded on the second
occurance.