New feature, not new line option though. Now shows init type with -x (also shows rc type if openrc). -xx shows init / rc version number. Change runlevel to target if systemd and if non numeric runlevel given. Should support systemd/upstart/epoch/runit sysvinit. Supports openrc as extra data if it's present. Rearranged -I line a bit but really just exchanged Runlevel: for Init: v: Runlevel: default:
This is the first step, some of the init system ID methods are weak and non robust
and this may need to be revised, but it should for now identify systemd/upstart quite
accurately, and in most cases sysvinit. Note that to get sysvinit version number requires
tool: strings which in debian/ubuntu is in package binutils. I don't know the package names
for arch/fedora/etc for the recommends check tool in inxi yet.

I believe this will be good enough for a first draft version, but over time we'll get it
more fine tuned, but as it is now, it should cover at least 99% of users, which isn't bad.